python/requirements.txt
customtkinter>=5.2.1
Pillow>=10.0.0
opencv-python>=4.8.0
face-recognition>=1.3.0
websocket-client>=1.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0 